MarkovChains

MarkovChains.ContinuousTimeMarkovChain

MarkovChains.ItoDiffusionProcess

MarkovChains.random_sample

MarkovChains.stationary_distribution

MarkovChains.ContinuousTimeMarkovChain — Type

ContinuousTimeMarkovChain{T}

Type for storing a continuous time Markov chain.

Can be constructed directly by supplying the infinitesimal generator infinitesimal_generator, or by supplying an ItoDiffusionProcess to be approximated.

See also https://en.wikipedia.org/wiki/Continuous-timeMarkovchain#Stationary_distribution

MarkovChains.ItoDiffusionProcess — Type

ItoDiffusionProcess

Type for storing an Itô diffusion process.

It consists of a function μ(x) for the drift term, and a function σ(x) for the diffusion term.

See also https://en.wikipedia.org/wiki/It%C3%B4_diffusion

MarkovChains.random_sample — Function

random_sample(mc::ContinuousTimeMarkovChain, Is::Integer=1, draws::Real=1000)

Given a starting index Is of type Integer, it returns a sequence of draws of length N.

The output Tuple contains a cumulative elapsed time T and the pseudorandom sequence of indices.

MarkovChains.stationary_distribution — Method

stationary_distribution(mc::ContinuousTimeMarkovChain, tol::T)

Computes the stationary distribution of the continuous time Markov chain mc.

By convention, the stationary distribution is treated as a PMF.

See also https://en.wikipedia.org/wiki/Continuous-timeMarkovchain#Stationary_distribution